backend/requirements.txt
psycopg2>=2.8
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.0,<5.0
django-cors-headers>=3.10,<4.0
pyotp>=2.6,<3.0
qrcode>=6.1,<7.0
django-oauth-toolkit>=2.4,<3.0
channels>=3.0,<4.0
python-decouple>=3.5,<4.0
Pillow>=9.0.0
channels-redis>=4.2.0